- Build fixes for new boost and vs 2012 - Remove global Job lock from Proactor - Add taglib src

+++ src/IO/src/PipeConnection.cpp
@@ -72,7 +72,7 @@ PipeConnection::beginWrite(std::istream& dataStream, boost::any tag /*= NULL*/)
 void 
 PipeConnection::beginWrite(const std::string& str, boost::any tag /*= NULL*/)
 {
-    unsigned int len = str.size();
+    size_t len = str.size();
     boost::shared_array<char> data(new char[len]);
     str.copy(data.get(), len);
     Proactor::instance()->beginWrite(this->shared_from_this(), boost::bind(&Connection::onWriteComplete,this,_1,_2), Buffer<char>(data, len, len), tag);
